[SERVICE-822] Move use of config store in FinAppDirectoryStorage
