+++ src/provider/model/AppDirectory.ts
@@ -7,10 +7,8 @@ import {Application, AppName, AppDirIntent} from '../../client/directory';
 import {AsyncInit} from '../controller/AsyncInit';
 import {CustomConfigFields} from '../constants';
 import {checkCustomConfigField, deduplicate} from '../utils/helpers';
-import {StoredAppDirectoryShard} from '../../client/internal';
 
-import {ConfigStoreBinding} from './ConfigStore';
-import {AppDirectoryStorage, DomainAppDirectoryShard} from './AppDirectoryStorage';
+import {AppDirectoryStorage, ShardScope} from './AppDirectoryStorage';
 
 enum StorageKeys {
     DIRECTORY_CACHE = 'fdc3@directoryCache'
@@ -19,22 +17,6 @@ enum StorageKeys {
 interface CacheEntry {
     url: string;
     applications: Application[];
-}
-
-interface GlobalShardScope {
-    type: 'global';
-}
-
-interface DomainShardScope {
-    type: 'domain';
-    domain: string;
-}
-
-type ShardScope = GlobalShardScope | DomainShardScope;
-
-interface ScopedAppDirectoryShard {
-    scope: ShardScope;
-    shard: StoredAppDirectoryShard;
 }
 
 @injectable()
@@ -85,20 +67,15 @@ export class AppDirectory extends AsyncInit {
     public readonly directoryChanged: Signal<[]> = new Signal();
 
     private readonly _appDirectoryStorage: AppDirectoryStorage;
-    private readonly _configStore: ConfigStoreBinding;
 
     private readonly _fetchedUrls: Set<string> = new Set();
 
     private _directory: Application[] = [];
 
-    public constructor(
-        @inject(Inject.APP_DIRECTORY_STORAGE) appDirectoryStorage: AppDirectoryStorage,
-        @inject(Inject.CONFIG_STORE) configStore: ConfigStoreBinding
-    ) {
+    public constructor(@inject(Inject.APP_DIRECTORY_STORAGE) appDirectoryStorage: AppDirectoryStorage) {
         super();
 
         this._appDirectoryStorage = appDirectoryStorage;
-        this._configStore = configStore;
     }
 
     public getAppByName(name: AppName): Promise<Application | null> {
@@ -121,7 +98,6 @@ export class AppDirectory extends AsyncInit {
     protected async init(): Promise<void> {
         this._appDirectoryStorage.changed.add(this.onStorageChanged, this);
 
-        await this._configStore.initialized;
         await this.refreshDirectory();
     }
 
@@ -132,26 +108,7 @@ export class AppDirectory extends AsyncInit {
     }
 
     private async refreshDirectory(): Promise<void> {
-        const configUrl = this._configStore.config.query({level: 'desktop'}).applicationDirectory;
-
-        const scopedShards: ScopedAppDirectoryShard[] = [
-            {
-                scope: {
-                    type: 'global'
-                },
-                shard: {
-                    urls: configUrl ? [configUrl] : [],
-                    applications: []
-                }
-            },
-            ...this._appDirectoryStorage.getDirectoryShards().map((shard: DomainAppDirectoryShard) => ({
-                scope: {
-                    type: 'domain',
-                    domain: shard.domain
-                } as ShardScope,
-                shard: shard.shard
-            }))
-        ];
+        const scopedShards = this._appDirectoryStorage.getDirectoryShards();
 
         const remoteDirectorySnippets = await parallelMap(scopedShards, async (scopedShard) => {
             return parallelMap(filterUrlsByScope(scopedShard.scope, scopedShard.shard.urls), async (url) => {
