+++ src/provider/model/AppDirectoryStorage.ts
@@ -2,13 +2,24 @@ import {Signal} from 'openfin-service-signal';
 
 import {StoredAppDirectoryShard} from '../../client/internal';
 
-export interface DomainAppDirectoryShard {
+export interface GlobalShardScope {
+    type: 'global';
+}
+
+export interface DomainShardScope {
+    type: 'domain';
     domain: string;
+}
+
+export type ShardScope = GlobalShardScope | DomainShardScope;
+
+export interface ScopedAppDirectoryShard {
+    scope: ShardScope;
     shard: StoredAppDirectoryShard;
 }
 
 export interface AppDirectoryStorage {
     readonly changed: Signal<[]>;
 
-    getDirectoryShards(): DomainAppDirectoryShard[];
+    getDirectoryShards(): ScopedAppDirectoryShard[];
 }
